refactor: update import path for report-issue schema for consistency
